Update the client & server files with logic for sending & receiving chat messages.
In client side (main.js), "submit" event occurs when a user hits the submit button after typing out the message.
chatForm.addEventListener() listens for this 'submit' event, extracts the message from DOM and sends it to server as an event called "chatMessage" (using socket.emit() ).

In the server side (app.js), listen for an event called 'chatMessage' and when that event occurs, broadcast it as a 'message' event to all clients (including the one that sent the message) .

In the client side, keep an event listener for 'messsage' event which when encountered will take up the message and put it in the chat room via DOM Manipulation. This DOM Manipulation is done by outputMsgToDOM()