Add a separate utils file for all user related actions like creating a user, storing their data, finding a user etc.
File is './utils/users.js' - in this we have an array of user objects. A user object consists of 3 things viz. id (=socket.id), username & the chat room.

userJoin() is called in app.js when a new user joins a particular room
getCurrentUser() returns the user object if id matches with a stored object in the object array